Handle review comments
