+++ fcli-core/fcli-fod/src/main/java/com/fortify/cli/fod/_common/session/cli/cmd/FoDSessionLoginCommand.java
@@ -13,7 +13,9 @@
 package com.fortify.cli.fod._common.session.cli.cmd;
 
 import com.fortify.cli.common.exception.FcliSimpleException;
+import com.fortify.cli.common.exception.FcliTechnicalException;
 import com.fortify.cli.common.output.cli.mixin.OutputHelperMixins;
+import com.fortify.cli.common.rest.unirest.UnexpectedHttpResponseException;
 import com.fortify.cli.common.rest.unirest.config.IUrlConfig;
 import com.fortify.cli.common.session.cli.cmd.AbstractSessionLoginCommand;
 import com.fortify.cli.common.session.cli.mixin.ISessionNameSupplier;
@@ -23,6 +25,8 @@ import com.fortify.cli.fod._common.session.helper.FoDSessionDescriptor;
 import com.fortify.cli.fod._common.session.helper.FoDSessionHelper;
 import com.fortify.cli.fod._common.session.helper.oauth.FoDOAuthHelper;
 import com.fortify.cli.fod._common.session.helper.oauth.FoDTokenCreateResponse;
+import com.fortify.cli.fod._common.session.helper.oauth.impl.BasicFoDUserAuthCode;
+import com.fortify.cli.fod._common.session.helper.oauth.impl.BasicFoDUserCredentials;
 
 import lombok.Getter;
 import picocli.CommandLine.Command;
@@ -34,12 +38,12 @@ public class FoDSessionLoginCommand extends AbstractSessionLoginCommand<FoDSessi
     @Getter private FoDSessionHelper sessionHelper = FoDSessionHelper.instance();
     @Mixin private FoDSessionLoginOptions loginOptions;
     @Mixin private FoDUnirestInstanceSupplierMixin unirestInstanceSupplierMixin;
-    
+
     @Override
     public ISessionNameSupplier getSessionNameSupplier() {
         return unirestInstanceSupplierMixin;
     }
-    
+
     @Override
     protected void logoutBeforeNewLogin(String sessionName, FoDSessionDescriptor sessionDescriptor) {
         unirestInstanceSupplierMixin.close(sessionName);
@@ -50,12 +54,43 @@ public class FoDSessionLoginCommand extends AbstractSessionLoginCommand<FoDSessi
     protected FoDSessionDescriptor login(String sessionName) {
         FoDSessionDescriptor sessionDescriptor;
         IUrlConfig urlConfig = loginOptions.getUrlConfigOptions();
-        if ( loginOptions.hasClientCredentials() ) {
+        if (loginOptions.hasClientCredentials()) {
             FoDTokenCreateResponse createTokenResponse = FoDOAuthHelper.createToken(urlConfig, loginOptions.getClientCredentialOptions(), loginOptions.getAuthOptions().getScopes());
             sessionDescriptor = new FoDSessionDescriptor(urlConfig, createTokenResponse);
-        } else if ( loginOptions.hasUserCredentials() ) {
-            FoDTokenCreateResponse createTokenResponse = FoDOAuthHelper.createUserToken(urlConfig, loginOptions);
-            sessionDescriptor = new FoDSessionDescriptor(urlConfig, createTokenResponse);
+        } else if (loginOptions.hasUserCredentials()) {
+            var credBuilder = BasicFoDUserCredentials.builder()
+                    .tenant(loginOptions.getUserCredentialOptions().getTenant())
+                    .user(loginOptions.getUserCredentialOptions().getUser())
+                    .password(loginOptions.getUserCredentialOptions().getPassword());
+
+            var authCodeBuilder = BasicFoDUserAuthCode.builder()
+                    .securityCode(loginOptions.getUserCredentialOptions().getSecurityCode())
+                    .isTotp(loginOptions.getUserCredentialOptions().isTotp());
+            try {
+                FoDTokenCreateResponse createTokenResponse = FoDOAuthHelper.createToken(urlConfig, credBuilder.build(),
+                        authCodeBuilder.build(), loginOptions.getAuthOptions().getScopes());
+                sessionDescriptor = new FoDSessionDescriptor(urlConfig, createTokenResponse);
+            } catch (UnexpectedHttpResponseException e) {
+                if (e.getStatus() == 400) {
+                    String mfaGuidance = "If MFA is required, provide the security code:\n"
+                            + "  --code <code>  (or -c <code>) to provide the security code\n"
+                            + "  --totp          to indicate the code is from a TOTP authenticator app";
+
+                    String errorWithCode = "Authentication failed. Possible causes:\n"
+                            + "  - Incorrect username or password\n"
+                            + "  - MFA/TOTP code incorrect, expired, or wrong type (TOTP vs MFA)\n"
+                            + "Please verify your credentials and MFA/TOTP code if applicable:\n"
+                            + mfaGuidance;
+
+                    String errorWithoutCode = "Authentication failed. Possible causes:\n" +
+                            "  - Incorrect username or password\n" +
+                            "  - FoD tenant requires MFA/TOTP authentication\n\n" +
+                            mfaGuidance;
+                    String errorMessage = loginOptions.hasSecurityCode() ? errorWithCode : errorWithoutCode;
+                    throw new FcliSimpleException(errorMessage);
+                }
+                throw new FcliTechnicalException(e.getMessage(), e);
+            }
         } else {
             throw new FcliSimpleException("Either FoD client or user credentials must be provided");
         }
